Fix Apple Clang warnings (cherry-picked from wip-ios-support)
(cherry picked from commit 983ee97746dcfadea8eccf57665f3b6eb9d456b1)

+++ src/control/controls.cpp
@@ -1245,7 +1245,7 @@ void RemoveNullInputMethods()
         if(Controls::g_InputMethods[i])
             i++;
         else
-            Controls::DeleteInputMethodSlot(i);
+            Controls::DeleteInputMethodSlot((int)i);
     }
 }
